define [[provides]] in build plan
Ensure that the buildpack defines what it provides so that other buildpacks that depend on this buildpack can specify that requirement. The reason these two were chosen was to align with the name that the riff java buildpack uses to require `openjdk-jre` and `jvm-application`.

I don't know that this is where you would want to output the `provides`. I don't know enough about the buildpack to know if it _sometimes_ provides these things and other times it does not. If you want, we could pair on this. I just wanted somewhere to have the discussion.